pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "wbb-demo"
version = "0.1.0"
description = "Neuro-symbolic intrinsic value calibration (Fi + Fm) with LLM semantic modulators"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pydantic>=2.5",
  "torch>=2.0",
  "yfinance>=0.2.40",
  "matplotlib>=3.8",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
openai = ["openai>=1.0"]
gemini = [
  "google-genai>=1.0.0",
  "google-generativeai>=0.8.0",
  "python-dotenv>=1.0.0",
]
pipeline = [
  "yfinance>=0.2.40",
  "matplotlib>=3.8",
  "requests>=2.31",
]
macro = [
  "akshare>=1.14.0",
]
moutai = [
  "pdfplumber>=0.11",
  "akshare>=1.14.0",
  "python-docx>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["value_investment_agent*"]
